# v2/api/openapi/public.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "agentserver v2 user resource API",
    "version": "0.1.0",
    "description": "Platform and Browser user resources. Core requires the matching gateway workload identity and validates every original opaque user bearer through Hydra introspection, including its exact client, audience, action scopes, and workspace resource grant."
  },
  "servers": [{"url": "https://agentserver-core"}],
  "security": [{"browserGatewayMTLS": [], "browserOAuth": ["runs:read"]}],
  "paths": {
...
    "/v2/workspaces/{workspaceId}/executors": {
      "parameters": [
        {"$ref": "#/components/parameters/WorkspaceId"}
      ],
      "get": {
        "operationId": "listExecutorResources",
        "description": "Lists bounded executor state for one active workspace.",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["executors:read"]}],
        "responses": {
          "200": {"description": "Workspace executors", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListExecutorResourcesResponse"}}}},
          "401": {"$ref": "#/components/responses/PublicError"},
          "403": {"$ref": "#/components/responses/PublicError"},
          "404": {"$ref": "#/components/responses/PublicError"}
        }
      },
      "post": {
        "operationId": "createExecutorResource",
        "description": "Creates one enrolling executor identity. Core rechecks that the current user is an owner of the active workspace.",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["executors:create"]}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateExecutorResourceRequest"}}}},
        "responses": {
          "200": {"description": "Exact executor identity already exists", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateExecutorResourceResponse"}}}},
          "201": {"description": "Executor identity created in enrolling state", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateExecutorResourceResponse"}}}},
          "400": {"$ref": "#/components/responses/PublicError"},
          "401": {"$ref": "#/components/responses/PublicError"},
          "403": {"$ref": "#/components/responses/PublicError"},
          "404": {"$ref": "#/components/responses/PublicError"},
          "409": {"$ref": "#/components/responses/PublicError"}
        }
      }
    },
    "/v2/workspaces/{workspaceId}/executors/{executorId}": {
      "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/ExecutorId"}],
      "delete": {
        "operationId": "archiveExecutorResource",
        "description": "Owner-only executor revocation that fences connections, disables environments, and revokes pending enrollment tokens atomically.",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["executors:archive"]}],
        "responses": {
          "200": {"description": "Revoked executor", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ArchiveExecutorResourceResponse"}}}},
          "400": {"$ref": "#/components/responses/PublicError"},
          "401": {"$ref": "#/components/responses/PublicError"},
          "403": {"$ref": "#/components/responses/PublicError"},
          "404": {"$ref": "#/components/responses/PublicError"},
          "409": {"$ref": "#/components/responses/PublicError"}
        }
      }
    },
    "/v2/workspaces/{workspaceId}/executors/{executorId}:enrollmentToken": {
      "parameters": [
        {"$ref": "#/components/parameters/WorkspaceId"},
        {"$ref": "#/components/parameters/ExecutorId"}
      ],
      "post": {
        "operationId": "issueExecutorEnrollmentToken",
        "description": "Issues or exactly recovers one short-lived, single-executor enrollment bearer. A new idempotency key revokes the prior unconsumed bearer authority.",
        "security": [{"platformGatewayMTLS": [], "platformOAuth": ["executors:enroll"]}],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "responses": {
          "200": {"description": "Exact idempotent token request recovered", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IssueExecutorEnrollmentTokenResponse"}}}},
          "201": {"description": "Enrollment token issued", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IssueExecutorEnrollmentTokenResponse"}}}},
          "400": {"$ref": "#/components/responses/PublicError"},
          "401": {"$ref": "#/components/responses/PublicError"},
          "403": {"$ref": "#/components/responses/PublicError"},
          "404": {"$ref": "#/components/responses/PublicError"},
          "409": {"$ref": "#/components/responses/PublicError"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "browserGatewayMTLS": {"type": "mutualTLS", "description": "Exact configured browser-gateway SPIFFE workload identity."},
      "platformGatewayMTLS": {"type": "mutualTLS", "description": "Exact configured platform-gateway SPIFFE workload identity."},
      "browserOAuth": {
        "type": "oauth2",
        "description": "Original agentserver-browser token. Core requires active introspection, exact client_id, single aud=agentserver-browser-api, and one workspace resource grant matching the request path.",
        "flows": {"authorizationCode": {"authorizationUrl": "https://auth-sg.byted.bps.dev/oauth2/auth", "tokenUrl": "https://auth-sg.byted.bps.dev/oauth2/token", "scopes": {
          "sessions:read": "List and read the current user's sessions",
          "sessions:create": "Create a session for the current user",
          "sessions:update": "Rename the current user's session",
          "sessions:archive": "Archive the current user's session",
          "runs:read": "Read the current user's run state and committed event projections",
          "runs:create": "Create a run in the current user's session",
          "runs:cancel": "Cancel the current user's run",
          "approvals:decide": "Approve or deny a live approval for the current user's run"
        }}}
      },
      "platformOAuth": {
        "type": "oauth2",
        "description": "Original agentserver-platform token. Core requires active introspection, exact client_id, single aud=agentserver-platform-api, and a matching global or workspace management grant.",
        "flows": {"authorizationCode": {"authorizationUrl": "https://auth-sg.byted.bps.dev/oauth2/auth", "tokenUrl": "https://auth-sg.byted.bps.dev/oauth2/token", "scopes": {
          "workspaces:read": "List and read granted workspaces",
          "workspaces:create": "Create a workspace",
          "workspaces:update": "Update a granted workspace",
          "workspaces:archive": "Archive a granted workspace",
          "members:read": "Read workspace members",
          "members:add": "Add a workspace member",
          "members:update": "Change a workspace member role",
          "members:remove": "Remove a workspace member",
          "executors:read": "Read workspace executor state",
          "executors:create": "Create a workspace executor identity",
          "executors:enroll": "Issue a workspace executor enrollment token",
          "executors:update": "Update workspace executor metadata",
          "executors:archive": "Archive a workspace executor",
          "llm-gateways:read": "Read workspace LLM Gateway configuration and the current user's grant status",
          "llm-gateways:create": "Create a workspace LLM Gateway",
          "llm-gateways:disable": "Disable a workspace LLM Gateway",
          "llm-gateway-grants:authorize": "Authorize the current user's third-party Gateway grant",
          "llm-gateway-grants:revoke": "Revoke the current user's third-party Gateway grant"
        }}}
      }
...
    "parameters": {
...
      "ExecutorId": {"name": "executorId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/UUID"}},
...
      "WorkspaceId": {"name": "workspaceId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/UUID"}},
...
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 256, "pattern": "^[!-~]+$"}}
    },
    "responses": {
      "PublicError": {"description": "Bounded public error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicErrorResponse"}}}}
    },
    "schemas": {
      "UUID": {"type": "string", "format": "uuid", "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$", "not": {"const": "00000000-0000-0000-0000-000000000000"}},
...
      "CreateExecutorResourceRequest": {
        "type": "object", "additionalProperties": false, "required": ["executorId"],
        "properties": {
          "executorId": {"$ref": "#/components/schemas/UUID"}
        }
      },
      "ExecutorResourceState": {
        "type": "object", "additionalProperties": false,
        "required": ["executorId", "workspaceId", "status", "version", "createdAt", "updatedAt"],
        "properties": {
          "executorId": {"$ref": "#/components/schemas/UUID"},
          "workspaceId": {"$ref": "#/components/schemas/UUID"},
          "status": {"enum": ["enrolling", "offline", "online", "revoked"]},
          "version": {"type": "integer", "minimum": 1, "maximum": 9007199254740991},
          "createdAt": {"type": "string", "format": "date-time"},
          "updatedAt": {"type": "string", "format": "date-time"}
        }
      },
      "CreateExecutorResourceResponse": {
        "type": "object", "additionalProperties": false, "required": ["executor", "created"],
        "properties": {
          "executor": {"$ref": "#/components/schemas/ExecutorResourceState"},
          "created": {"type": "boolean"}
        }
      },
      "ListExecutorResourcesResponse": {
        "type": "object", "additionalProperties": false, "required": ["executors"],
        "properties": {"executors": {"type": "array", "maxItems": 256, "items": {"$ref": "#/components/schemas/ExecutorResourceState"}}}
      },
      "ArchiveExecutorResourceResponse": {
        "type": "object", "additionalProperties": false, "required": ["executor", "changed"],
        "properties": {"executor": {"$ref": "#/components/schemas/ExecutorResourceState"}, "changed": {"type": "boolean"}}
      },
      "IssueExecutorEnrollmentTokenResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["executorId", "token", "expiresAt", "created"],
        "properties": {
          "executorId": {"$ref": "#/components/schemas/UUID"},
          "token": {"type": "string", "minLength": 1, "maxLength": 8192, "pattern": "^asv2enr1\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+$", "description": "Sensitive short-lived enrollment bearer; clients must not log or persist it", "x-agentserver-sensitive": true},
          "expiresAt": {"type": "string", "format": "date-time"},
          "created": {"type": "boolean"}
        }
      },
...
      "PublicErrorResponse": {
        "type": "object", "additionalProperties": false, "required": ["code", "message"],
        "properties": {
          "code": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[a-z0-9_]+$"},
          "message": {"type": "string", "minLength": 1, "maxLength": 1024},
          "currentRunId": {"$ref": "#/components/schemas/UUID"}
        }
      }
    }
  },
...
}